requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow
uvicorn
